Renamed unit conversion utility
